Added api endpoints for reports and associated files
This commit adds api end points for reports and associated files so
that they can be filtered and downloaded by client applications. It
also include small tweaks to the project filters and associated
templates to ensure consistent use of filters across
applications. Small updates to the tests were also required to get the
me to pass again.  Finally, this commit add drf-yasg to the
requirements and uses it to create a swagger json api endpoint to
access the available filters. All tests continue to pass.