Blackened and pysorted several files. Removed xfail markers
This commit is a simple clean up commit that formats several files,
removed unused imports, and applied pysort for consistency. There were
several tests marks with xfail that were actually passing. The xfail
marker was removed. Currently 410 tests are passing.  There are two
that continue to be marked with xfail and one that is being skipped.