Updated choice fields to use list instead of a dictionary.
This commit fixes a bug in how the choice fields were specified - they
must be specified with a list or tuple not a set (with curly braces).
Specifying choice options as a set produces random, superfluous
migrations as the order of a set can change.  These in-turn produce
hard to debug migration conflicts.  This commit replaces all of sets
with lists. Hopefully it won't cause any further issues in the
future...